+++ CHANGELOG.md
@@ -18,7 +18,7 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Fixed
 
-- [#11](https://github.com/zendframework/ZendXml/pull/11) updates the
+- [zendframework/zendxml#11](https://github.com/zendframework/ZendXml/pull/11) updates the
   dependencies to PHP `^5.3.3 || ^7.0` and PHPUnit `^3.7 || ^4.0`, ensuring
   better compatibility with other components, and with PHP 7. The test matrix
   was also expanded to add PHP 7 as a required platform.